move SwiftUI Extension to Asset File
